refactor(layout): Update and clean Rout Error Boundary
ref: #232 #174 #230
